requirements.txt
pysqlite3-binary
langchain-chroma>=0.1.2
langchain-google-genai
langchain-huggingface
sentence-transformers
pypdf
python-dotenv
streamlit
langchain
langchain-community
langchain-text-splitters